Evaluate init clauses in loop body if appropriate (#73)
This patch moves the evaluation of the `init` clause of at-reduce
macro to the loop body, instead of using `Transducers.OnInit`.  This
lets us write loops relying on the first element for initializing the
accumulator. For example, `zero(x)` is the `init` clause in the
following code below and it accesses the value `x` only available
inside the loop:

[at]floop ex for x in arrays
[at]reduce() do (s = zero(x); x)
s .+= x
end
end

When the `init` clauses do not access the variables assigned inside
the loop, they are hoisted out, to improve type inference. For example,
when `[at]reduce` is written after `continue`,

next(rf, start(rf, init), first(xs))

would be inferred to contain `_FLoopInit` (since `[at]reduce` may not
be invoked) if we chose to invoke the `init` clauses always inside the
loop body.  Instead, if we invoke the `init` clauses inside `start`,
we can return non-`_FLoopInit` accumulator from `start(rf, init)` and
it helps avoiding this kind of type instability.